pyproject.toml
[project]
name = "scheduleremailnotification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "celery>=5.6.3",
    "fastapi>=0.136.0",
    "jinja2>=3.1.6",
    "asyncpg>=0.30.0",
    "fastapi>=0.136.0",
    "pydantic-settings>=2.11.0",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.46.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.11",
]

[tool.ruff]
line-length = 88
target-version = "py310"
fix = true
